# Optimize V1 execution memoization