--- a/.changeset/new-actors-breathe.md
+++ b/.changeset/new-actors-breathe.md
@@ -1,0 +1,5 @@
+---
+"inngest": patch
+---
+
+Optimize function memoization