--- a/packages/inngest/src/helpers/promises.ts
+++ b/packages/inngest/src/helpers/promises.ts
@@ -35,12 +35,12 @@ export const createFrozenPromise = (): Promise<unknown> => {
  * Returns a Promise that resolves after the current event loop's microtasks
  * have finished, but before the next event loop tick.
  */
-export const resolveAfterPending = (count = 1000): Promise<void> => {
+export const resolveAfterPending = (count = 10): Promise<void> => {
   /**
    * This uses a brute force implementation that will continue to enqueue
-   * microtasks 1000 times before resolving. This is to ensure that the
-   * microtask queue is drained, even if the microtask queue is being
-   * manipulated by other code.
+   * microtasks 10 times before resolving. This is to ensure that the microtask
+   * queue is drained, even if the microtask queue is being manipulated by other
+   * code.
    *
    * While this still doesn't guarantee that the microtask queue is drained,
    * it's our best bet for giving other non-controlled promises a chance to