--- a/packages/inngest/test/benchmark/execution/util.ts
+++ b/packages/inngest/test/benchmark/execution/util.ts
@@ -49,7 +49,7 @@ export const createExecutionWithMemoizedSteps = ({
     const execution = client
       .createFunction({ id: "test" }, { event: "test" }, async ({ step }) => {
         for (let i = 0; i < stepCount; i++) {
-          await step.run(userStepId, () => userStepOutput);
+          await step.run(i === 0 ? userStepId : `${userStepId}${STEP_INDEXING_SUFFIX}${i}`, () => userStepOutput);
         }
       })
       ["createExecution"]({